Migrate from kingpin to cobra
