+++ go.mod
@@ -19,8 +19,12 @@ require (
 	gopkg.in/yaml.v2 v2.4.0
 )
 
+require github.com/spf13/cobra v1.7.0
+
 require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/rivo/uniseg v0.2.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/tkuchiki/go-timezone v0.2.2 // indirect
 )
